2025_fall/doom_3d/requirements.txt
pygame
PyOpenGL
PyOpenGL-accelerate
numpy
Pillow
